+++ code/src/cljs/sixsq/nuvla/ui/dashboard/utils.cljs
@@ -34,21 +34,6 @@
   (assoc evt :delta-time (time/delta-minutes start end)))
 
 
-(defn has-action?
-  [action deployment]
-  (->> deployment
-       :operations
-       (filter #(= action (:rel %)))
-       not-empty
-       boolean))
-
-
-(def stop-action? (partial has-action? "stop"))
-
-
-(def delete-action? (partial has-action? "delete"))
-
-
 (defn is-started?
   [state]
   (= state "STARTED"))