+++ code/src/cljs/sixsq/nuvla/ui/welcome/views.cljs
@@ -56,6 +56,6 @@
       (when-not @iframe? [card :application :welcome-application-desc "play" "apps"])
       [card :data :welcome-data-desc "database" "data"]
       [card :deployment :welcome-deployment-desc "cloud" "deployment"]
-      (when-not @iframe? [card :nuvlabox-ctrl :welcome-nuvlabox-desc "desktop" "nuvlabox"])
+      (when-not @iframe? [card :nuvlabox-ctrl :welcome-nuvlabox-desc "desktop" "edge"])
       (when-not @iframe? [card :api :welcome-api-desc "code" "api"])
       ]]))